+++ frontend/src/pages/ErrorsV2/SearchPanel/SearchPanel.tsx
@@ -16,7 +16,7 @@ import { ErrorFeedCard } from '@pages/ErrorsV2/ErrorFeedCard/ErrorFeedCard'
 import ErrorFeedHistogram from '@pages/ErrorsV2/ErrorFeedHistogram/ErrorFeedHistogram'
 import ErrorQueryBuilder from '@pages/ErrorsV2/ErrorQueryBuilder/ErrorQueryBuilder'
 import { useGlobalContext } from '@routers/OrgRouter/context/GlobalContext'
-import { gqlSanitize } from '@util/gqlSanitize'
+import { gqlSanitize } from '@util/gql'
 import { useParams } from '@util/react-router/useParams'
 import clsx from 'clsx'
 import { useEffect, useState } from 'react'
